+++ cli/provider/cmd.go
@@ -573,12 +573,14 @@ func (c *CmdConfigurator) ValidateFlags(ctx context.Context, cmd *cobra.Command)
 			c.cfg.Path = utils.ToAbsPath(c.logger, path)
 			return nil
 		}
+
 		// handle the app command
 		if c.cfg.Command == "" {
 			if !alreadyRunning(cmd.Name(), c.cfg.Test.BasePath) {
 				return c.noCommandError()
 			}
 		}
+		
 		// set the command type
 		c.cfg.CommandType = string(utils.FindDockerCmd(c.cfg.Command))
